fix(build): adjust the release script to create valid binaries
Two things broke the release binaries that were not accounted for in its first
form. First, the `RUSTFLAGS` of the host OS were used during the compilation
which resulted in invalid binaries that can't be run on different systems.

Second, for some reason the archives compressed by `tar` are apparently not
valid. So instead we have to create plain tar archives and then compress
ourselves.